fix: [rhui4] run 'rhui-manager repo unused --by_repo_id' synchronous way
rhua must pull the entitlement certification before its administrator
tries to add a repository.

In RHUI3, there was no issue. We uploaded the entitlement
certification manually downloaded from access.redhat.com to rhui in
the 80.sh. When adding a repository, the certification was in rhui.

If an administrator uses TUI, there is no issue. When selecting a menu
entry for adding repositories/prodcuts, rhui-manager uploads an
entitlement certification IF rhui has no entitlement certification.

With RHUI4's rhui-manager in a shell script, there is an issue.  There
is no good timing to upload an entitlement certification. As the
result, adding operation in the shell script can fail. The original
script lists unused repositories before adding a repository.  It seems
that listing unused repositories uploads an entitlement
certification. At a glance, the adding operation works well.  However,
the "listing unused repositories" run in background, therefore we
cannot expect rhui has an entitlement certification when adding a
repository from the shell script. This change runs the "listing unused
repositories" in foreground. As a result, the adding operation works
well.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>